Applying force with care.
We only want to force an update of the thumb position when it is likely that the min/max values have changed.  Previously, the thumb value was always set to null, even if the slider hadn't been rendered yet.  In this case, the change event would be fired immediately upon rendering, and handleChange would end up setting the map zoom to zero even if that was not the initial value.  See the test (or the zoomslider.js) example for effects before and after.